% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_url.R
\name{construct_query_url}
\alias{construct_query_url}
\title{Build url query}
\usage{
construct_query_url(wh_state)
}
\arguments{
\item{wh_state}{State}
}
\value{
query url
}
\description{
Build url query
}
