% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uspvdb_data.R
\name{get_uspvdb_data}
\alias{get_uspvdb_data}
\title{Get USPVDB data}
\usage{
get_uspvdb_data(wh_state)
}
\arguments{
\item{wh_state}{State to get data for}
}
\value{
df Dataframe of requested data from USPVDB API
}
\description{
Get USPVDB data
}
